Extend CollectionType intead of extending Array
Extending Array was great, but it'd be nicer if we could extend _any_
collection type that contains Decodable objects. This `decode` function
will still always return an Array, for the same reason that the stdlib's
`map`, `flatMap`, etc return Arrays, but if we ever get Real Actual
Higher Kinded Types, we can fix this.